fix: fix bug for chosenChatDetails in redux
