+++ front-end/src/MainPage/Messenger/ChatRoom.js
@@ -1,3 +1,5 @@
+// 無用
+
 import React from 'react';
 import { styled } from '@mui/system';
 import { Typography } from '@mui/material';
