+++ front-end/src/MainPage/Messenger/MessageAreaHeader.js
@@ -1,3 +1,4 @@
+// 無用
 import React from 'react';
 import { styled } from '@mui/system';
 // 缺少avatar
