+++ front-end/src/MainPage/Messenger/MessageInputArea.js
@@ -1,3 +1,5 @@
+// 無用 => to MessageTextField.js
+
 import React, { useState } from 'react';
 import { styled } from '@mui/system';
 import { connect } from 'react-redux';
