+++ front-end/src/MainPage/Messenger/Messenger.js
@@ -1,8 +1,9 @@
+// 無用 => messageContetArea
+
 import React from 'react';
 import { styled } from '@mui/system';
 import { connect } from 'react-redux';
-import { HomepageMessage } from './HomepageMessage';
-import MessageAreaContent from './MessageAreaContent';
+import { MeesageIfNoChosenContact } from './MeesageIfNoChosenContact';
 const MessengerWrapper = styled('div')({
   // 空間足夠時允許擴展
   flexGrow: 1,
@@ -18,9 +19,9 @@ const Messenger = ({ chosenChatDetails }) => {
     // MessengerContent裡面的東西會根據chosenChatDetails來做相對應的找DB及渲染
     <MessengerWrapper>
       {!chosenChatDetails ? (
-        <HomepageMessage />
+        <MeesageIfNoChosenContact />
       ) : (
-        <MessageAreaContent chosenChatDetails={chosenChatDetails} />
+        <MeesageIfNoChosenContact chosenChatDetails={chosenChatDetails} />
       )}
     </MessengerWrapper>
   );
