--- front-end/src/MainPage/Messenger/SingleChatBubble.js
@@ -1,182 +1,0 @@
-import React from 'react';
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
-import Avatar from '@material-ui/core/Avatar';
-import { deepOrange } from '@material-ui/core/colors';
-import styled from '@emotion/styled';
-
-const ChatRoomMainWrapper = styled('div')({
-  width: '75%',
-  display: 'flex',
-  marginTop: '10px',
-});
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    messageRow: {
-      display: 'flex',
-    },
-    messageRowRight: {
-      display: 'flex',
-      justifyContent: 'flex-end',
-    },
-    messageBlue: {
-      position: 'relative',
-      marginLeft: '20px',
-      marginBottom: '10px',
-      padding: '10px',
-      backgroundColor: '#A8DDFD',
-      width: '60%',
-      //height: "50px",
-      textAlign: 'left',
-      font: "400 .9em 'Open Sans', sans-serif",
-      border: '1px solid #97C6E3',
-      borderRadius: '10px',
-      '&:after': {
-        content: "''",
-        position: 'absolute',
-        width: '0',
-        height: '0',
-        borderTop: '15px solid #A8DDFD',
-        borderLeft: '15px solid transparent',
-        borderRight: '15px solid transparent',
-        top: '0',
-        left: '-15px',
-      },
-      '&:before': {
-        content: "''",
-        position: 'absolute',
-        width: '0',
-        height: '0',
-        borderTop: '17px solid #97C6E3',
-        borderLeft: '16px solid transparent',
-        borderRight: '16px solid transparent',
-        top: '-1px',
-        left: '-17px',
-      },
-    },
-    messageOrange: {
-      position: 'relative',
-      marginRight: '20px',
-      marginBottom: '10px',
-      padding: '10px',
-      backgroundColor: '#f8e896',
-      width: '60%',
-      //height: "50px",
-      textAlign: 'left',
-      font: "400 .9em 'Open Sans', sans-serif",
-      border: '1px solid #dfd087',
-      borderRadius: '10px',
-      '&:after': {
-        content: "''",
-        position: 'absolute',
-        width: '0',
-        height: '0',
-        borderTop: '15px solid #f8e896',
-        borderLeft: '15px solid transparent',
-        borderRight: '15px solid transparent',
-        top: '0',
-        right: '-15px',
-      },
-      '&:before': {
-        content: "''",
-        position: 'absolute',
-        width: '0',
-        height: '0',
-        borderTop: '17px solid #dfd087',
-        borderLeft: '16px solid transparent',
-        borderRight: '16px solid transparent',
-        top: '-1px',
-        right: '-17px',
-      },
-    },
-
-    messageContent: {
-      padding: 0,
-      margin: 0,
-    },
-    messageTimeStampRight: {
-      position: 'absolute',
-      fontSize: '.85em',
-      fontWeight: '300',
-      marginTop: '10px',
-      bottom: '-3px',
-      right: '5px',
-    },
-
-    orange: {
-      color: theme.palette.getContrastText(deepOrange[500]),
-      backgroundColor: deepOrange[500],
-      width: theme.spacing(4),
-      height: theme.spacing(4),
-    },
-    avatarNothing: {
-      color: 'transparent',
-      backgroundColor: 'transparent',
-      width: theme.spacing(4),
-      height: theme.spacing(4),
-    },
-    displayName: {
-      marginLeft: '20px',
-    },
-  })
-);
-
-// 他人
-const MessageLeft = (message, fromMe, username, date, sameTime) => {
-  // const message = message ? message : 'no message';
-  // const date = date ? date : '';
-  const photoURL = photoURL ? photoURL : '../shared/images/fake_avatar.png';
-  // const userName = username ? userName : 'no username';
-  const classes = useStyles();
-  return (
-    <>
-      <div className={classes.messageRow}>
-        <Avatar
-          alt={username}
-          className={classes.orange}
-          src={photoURL}
-        ></Avatar>
-        <div>
-          <div className={classes.displayName}>{username}</div>
-          <div className={classes.messageBlue}>
-            <div>
-              <p className={classes.messageContent}>{message}</p>
-            </div>
-            <div className={classes.messageTimeStampRight}>{date}</div>
-          </div>
-        </div>
-      </div>
-    </>
-  );
-};
-// 自己
-const MessageRight = (message, fromMe, username, date, sameTime) => {
-  const classes = useStyles();
-  // const message = message ? message : 'no message';
-  // const date = date ? date : '';
-  return (
-    <div className={classes.messageRowRight}>
-      <div className={classes.messageOrange}>
-        <p className={classes.messageContent}>{message}</p>
-        <div className={classes.messageTimeStampRight}>{date}</div>
-      </div>
-    </div>
-  );
-};
-
-// add props to content
-export const SingleChatRoomBubble = ({
-  content,
-  fromMe,
-  username,
-  date,
-  sameTime,
-}) => {
-  // 上方用obj
-  // const { content, fromMe, username, date, sameTime } = obj;
-  // 如果訊息是從自己來的 (自己發送的訊息)
-  if (fromMe && sameTime) {
-    return MessageRight(content, fromMe, username, date, sameTime);
-  }
-  return MessageLeft(content, fromMe, username, date, sameTime);
-};
